+++ agent/CHANGELOG.rst
@@ -9,6 +9,7 @@ Unreleased
 * Improve function to get bookings sum from backend
 * Improve function to get used values from cluster
 * Updated reconciliation to update all features with one request
+* Fix RLM parser to handle multiple features in the same license server
 
 3.0.3 -- 2023-08-17
 -------------------